Initial commit
This is a template project for STM32F4xx with Debugging in RAM and FLASH support.
The project is set up for EM::BLOCKS (V2.30) with external GNU-ARM Compiler.
However it will also work with the integrated ARM-Compiler.

Setting DEBUG -> Debug in RAM
Setting RELEASE -> Debug or run in FLASH

To adapt the template project to your needs, edit

1. the HSE_VALUE with the oscillator frequency in file "stm32f4xx.h" example: 8000000 => 8MHz
2. system_stm32f4xx.c: typically only set PLL_M to the oscillator frequency in MHz (8 => 8MHz) eventually check and set PLL_N, PLL_P and PLL_Q and check SystemCoreClock
3. stm32f4xx_conf.h uncomment the peripheral files if additional peripheral drivers are needed
4. set STACK_SIZE in startup_stm32f4xx.c